requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-storage-blob==12.19.1
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.2
daphne==4.0.0
distro==1.9.0
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isodate==0.6.1
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
oauthlib==3.2.2
openai==1.25.2
packaging==24.0
paramiko==3.4.0
Pillow==10.0.0
portalocker==2.8.2
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
python-dotenv==1.0.1
pytz==2023.3
pywin32==306
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
storage==0.0.4.3
tqdm==4.66.4
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.2.1
whitenoise==6.6.0
zope.interface==6.0
